backendOK
BackEnd terminado. landingpage y home in process

+++ api/src/models/Country.js
@@ -29,16 +29,10 @@ module.exports = (sequelize) => {
 			type : DataTypes.STRING
 		},
 		area        : {
-			type : DataTypes.INTEGER
+			type : DataTypes.FLOAT
 		},
 		population  : {
 			type : DataTypes.INTEGER
-		}
-		// ,
-		// createdInDb : {
-		// 	type         : DataTypes.BOOLEAN,
-		// 	allowNull    : false,
-		// 	defaultValue : true
-		// }
+		}		
 	});
 };
